Pavel_Volya/requirements.txt
pyyaml
urllib3
pygame
mutagen 
vosk
SpeechRecognition
wave
requests
vosk_tts
opencv-contrib-python
thread
replicate
pydub